Create README.md
Created in 5 days and designed to be a program for non-technical users to be able to choose 2 .csv files to compare and strip information from. This information would then be placed into a new .csv file named by the user in a location of their choice. SimpleInstruction.md has more information on the exact specifications.